Adds norminette things..
